Major refactor ManyBodyOper and Oper
I. Bending the interpretation of const a bit

Firstly, ManyBodyOper, Oper, and Matrix can now all be used to modify
their referents even if they are const.  The main reason this is done is
to reduce the code duplication caused by the const + non-const versions,
at the cost of making the assignment-like operators somewhat weird (they
return a const reference).  For Matrix, this should be fine since the
"correct" const version is really "Matrix<const T>", not "const
Matrix<T> &".

Unfortunately, there is no such division for ManyBodyOper and Oper yet,
which means there is no way to protect the contents, nor is there a way
to construct either of them on a "const double *" without doing an
explicit cast.  Implementing this would requires some code duplication
and/or template hackery, which we want to avoid (we already did this for
Matrix and it's not pretty).

II. Shared offset arrays + construction on the fly

Oper and ManyBodyOper have been changed to use shared offsets arrays
instead of storing their own individual arrays of pointers.  This means
both Oper and ManyBodyOper (and also Matrix, which is still the same)
are now like slice references ("fat pointers") to the matrix.  In
particular, Oper is now constructed on the fly by indexing operations
such as ManyBodyOper::oper (and Matrix by Oper::operator[]).

This has the unfortunate side-effect of requiring ManyBodyOper/Oper to
be passed in by value, as the temporaries constructed on the fly don't
like to bind to lvalue references.  The alternative would be to pass in
by const lvalue reference, but that is rather confusing since it
suggests they are const.  So for now we use `const` on the toplevel to
indicate constness, even though putting qualifiers in those positions is
just snake oil.

I hope these the construction-on-the-fly approach does not impact
performance severely.

Note: The reason operator[] was not used for ManyBodyOper is because in
the future this might be refactored into an templated function to allow
for different Oper types for each OperKind.